backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-groq
langchain-huggingface
sentence-transformers
chromadb
pydantic
requests
